implement function to get current dots and return as array
The extension to the string class is used to quickly convert the input into a boolean value. This is likely not an optimal location for this but for now this will do. I am considering using a lambda expression instead, or simply adding a function to the server class to make this conversion.